+++ sources/MapWalls.hpp
@@ -13,7 +13,7 @@ class MapWalls : public MapTemplate {
         {
         }
         void render(gf::RenderWindow& renderer, int sizeUnit);
-        std::vector<gf::Vector2i> getVertices();
+        std::vector<gf::Vector2f> getVertices();
 
     private :
         gf::Image m_image;
